Fix rendering GIFs with a delay time of 0
The default for these gifs seems to be to just put them at 10. It also
appears as though any gifs with a speed lower than this might need to be
changed as a lot of renderers seem to have 10 as a minimum, but we'll
cross that bridge when we come to it.